Gave precedence to the /mnt/blue_hydra path
Gave precedence to the /mnt/blue_hydra path if it exists in the OS.  This is so that constant writing to a sqlite file won't wear out storage on Mini PCs like the Raspberry Pi and/or Odroid.  SD cards and eMMC modules are not known for their long term reliability.